third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-waiting-for-a-key.html

<!DOCTYPE html>
<html>
    <head>
        <title>Waiting for a key.</title>
        <script src="encrypted-media-utils.js"></script>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
    </head>
    <body>
        <video id="testVideo"></video>
        <div id="log"></div>
        <script>
            async_test(function(test)
            {
                var video = document.getElementById('testVideo');
                var mediaKeySession = null;
                var sessionReadyReceived = false;
                var encryptedEventCount = 0;

                var keyId = null;
                var rawKey = new Uint8Array([0xeb, 0xdd, 0x62, 0xf1, 0x68, 0x14, 0xd2, 0x7b,
                                             0x68, 0xef, 0x12, 0x2a, 0xfc, 0xe4, 0xae, 0x3c]);

                function onEncrypted(event)
                {
                    assert_equals(event.target, video);
                    assert_true(event instanceof window.MediaEncryptedEvent);
                    assert_equals(event.type, 'encrypted');

                    // The same decryption key is used by both the audio and
                    // the video streams so only create a session once. To
                    // avoid issues when comparing the expected.txt file
                    // (which logs the events in the order they occur), create
                    // the session on the second event. This also ensures we
                    // see both events.
                    if (++encryptedEventCount != 2)
                        return;

                    mediaKeySession = video.mediaKeys.createSession();
                    waitForEventAndRunStep('message', mediaKeySession, onMessage, test);
                    mediaKeySession.generateRequest(event.initDataType, event.initData).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                function onMessage(event)
                {
                    assert_true(event instanceof window.MediaKeyMessageEvent);
                    assert_equals(event.target, mediaKeySession);
                    assert_equals(event.type, 'message');

                    keyId = extractSingleKeyIdFromMessage(event.message);
                    setTimeout(verifyWaitingForKey, 250);
                }

                function verifyWaitingForKey()
                {
                    // Make sure the video is NOT paused and not progressing
                    // before the session is updated.
                    // TODO(xhwang): Check "waiting" event after it's
                    // implemented. See http://crbug.com/337975
                    assert_false(video.paused);
                    assert_equals(video.currentTime, 0);

                    updateSession();
                }

                function updateSession()
                {
                    assert_not_equals(keyId, null);
                    var jwkSet = stringToUint8Array(createJWKSet(createJWK(keyId, rawKey)));
                    mediaKeySession.update(jwkSet).then(function(result) {
                        sessionReadyReceived = true;
                    }).catch(function(error) {
                        forceTestFailureFromPromise(test, error);
                    });
                }

                function onPlaying(event)
                {
                    // Not using waitForEventAndRunStep() to avoid too many
                    // EVENT(onTimeUpdate) logs.
                    video.addEventListener('timeupdate', onTimeUpdate, true);
                }

                function onTimeUpdate(event)
                {
                    if (event.target.currentTime < 0.5 || !sessionReadyReceived)
                        return;

                    test.done();
                }

                MediaKeys.create('org.w3.clearkey').then(function(mediaKeys) {
                    waitForEventAndRunStep('encrypted', video, onEncrypted, test);
                    waitForEventAndRunStep('playing', video, onPlaying, test);
                    video.src = '../content/test-encrypted.webm';
                    return video.setMediaKeys(mediaKeys);
                }).then(function(result) {
                    video.play();
                }).catch(function(error) {
                    forceTestFailureFromPromise(test, error);
                });
            }, 'Waiting for a key.');
        </script>
    </body>
</html>
